BD1 hibernate. save y load..

+++ BD1Hibernate/src/hlantern/NivelHabilidadDO.java
@@ -68,7 +68,7 @@ public class NivelHabilidadDO {
 		return efectividad;
 	}
 
-	public void setEfectividad(int efectividad) {
+	public void setEfectividad(float efectividad) {
 		this.efectividad = efectividad;
 	}
